+++ pytest_lambda/__init__.py
@@ -3,3 +3,4 @@
 __version__ = '1.0.0'
 
 from .fixtures import *
+from .util import *